dismissedEmployee

+++ src/main/java/org/example/service/EmployeeService.java
@@ -204,12 +204,11 @@ public class EmployeeService {
                     employee.setStatus(GeneralStatus.BLOCK);
                     return employeeRepository.save(employee);
                 }).orElseThrow(() -> ExceptionUtil.throwNotFoundException("employee with this id does not exist!"));
-        tokenStoreRepository.findAll()
-                .forEach(tokenStore -> {
-                    if (tokenStore.getEmployeeId().equals(id)) {
-                        tokenStoreRepository.delete(tokenStore);
-                    }
-                });
+        for (TokenStore tokenStore : tokenStoreRepository.findAll()) {
+            if (tokenStore != null && tokenStore.getEmployeeId().equals(id)) {
+                tokenStoreRepository.delete(tokenStore);
+            }
+        }
     }
 
 }
